+++ ccmt-cff-rest-api/chalicelib/routes/schemaList.py
@@ -3,7 +3,7 @@ def schema_list(centerId):
     """List all schemas.
     Todo: schemas should belong to a user / center, so check permissions; for now, all schemas are public.
     """
-    # app.check_permissions('forms', 'ListForms')
+    # app.check_permissions('forms', 'SchemasList')
     schemas = TABLES.schemas.scan(
         ProjectionExpression = "id, version, #value.title, cff_permissions",
         ExpressionAttributeNames = {"#value": "value"}